chore: moved autofac builder to dedicated directory
